Switch to uuid4, instead of uuid1.
Greg had the nerve to find datasets with id conflicts. We do have more
than a hundred created within a second in the logs, but this is far
lower than the 2^14 per 100ns documented for uuid1 to have problems.

There's little downside to uuid4 for our usage, so we may as well
switch.