Fix codecov reporting
Codecov expects a full coverage.xml. The normal .coverage
file output by pytest-cov is insufficient.

This can either be generated with these extra args to pytest-cov,
or by using the standalone `coverage` tool, which we could do only
in CI, but I'm not sure it's available in GHA by default, and we're
running most things in docker anyway. It's messy, these seems reasonable.