Upgrade to ruamel.yaml 0.18+
The change that was catching us out is whether plain timestamps are
parsed to include a tzinfo (new) or not (old). This caused a lot of
tests to fail. I've updated all the failing tests to expect datetimes
WITH timezone, which will cause the tests to fail with old versions of
ruamel.

eo-datasets itself was already comptible, and produces identical files
with either version of ruamel.yaml.